+++ src/property/test/properties.component.spec.ts
@@ -47,15 +47,20 @@ class ThyPropertiesTestBasicComponent {
 @Component({
     selector: 'thy-properties-test-column',
     template: `
-        <thy-properties thyColumn="3">
+        <thy-properties #properties thyColumn="3">
             <thy-property-item thyLabelText="姓名">张萌</thy-property-item>
             <thy-property-item thyLabelText="年龄">24</thy-property-item>
             <thy-property-item thyLabelText="电话">18500010001</thy-property-item>
-            <thy-property-item thyLabelText="居住地址">北京市朝阳区十八里店小区26号10001</thy-property-item>
+            <thy-property-item thyLabelText="电话">18500010001</thy-property-item>
+            <thy-property-item thyLabelText="居住地址" [thySpan]="addressItemSpan">北京市朝阳区十八里店小区26号10001</thy-property-item>
         </thy-properties>
     `
 })
-class ThyPropertiesTestColumnComponent {}
+class ThyPropertiesTestColumnComponent {
+    @ViewChild('properties') propertiesComponent: ThyPropertiesComponent;
+
+    addressItemSpan = 2;
+}
 
 @NgModule({
     imports: [ThyPropertyModule, CommonModule],
@@ -106,13 +111,6 @@ describe(`thy-properties`, () => {
             expect(trs[3].nativeElement.innerText).toEqual(basicComponent.user.age.toString());
         });
 
-        it('should displayed input when age item hovered', () => {
-            const ageEditorElement = fixture.debugElement.query(By.css('.age-input')).parent;
-            dispatchMouseEvent(ageEditorElement.nativeElement, 'mouseenter');
-            fixture.detectChanges();
-            expect(ageEditorElement.nativeElement.parentNode.classList).toContain(itemContentEditingClass);
-        });
-
         it('should displayed input when age item clicked', () => {
             basicComponent.editTrigger = 'click';
             fixture.detectChanges();
@@ -122,15 +120,19 @@ describe(`thy-properties`, () => {
             expect(ageEditorElement.nativeElement.parentNode.classList).toContain(itemContentEditingClass);
         });
 
-        it('should set keep editing success', () => {
+        it('should add hover trigger class when thyEditTrigger is hover', () => {
+            basicComponent.editTrigger = 'hover';
+            fixture.detectChanges();
+            const element = fixture.debugElement.query(By.css('.thy-properties'));
+            expect(element.nativeElement.classList).toContain('thy-properties-edit-trigger-hover');
+        });
+
+        it('should set editing success', () => {
             const ageEditorElement = fixture.debugElement.query(By.css('.age-input')).parent;
-            dispatchMouseEvent(ageEditorElement.nativeElement, 'mouseenter');
-            basicComponent.editItemComponent.setKeepEditing(true);
-            dispatchMouseEvent(ageEditorElement.nativeElement, 'mouseleave');
+            basicComponent.editItemComponent.setEditing(true);
             fixture.detectChanges();
             expect(ageEditorElement.nativeElement.parentNode.classList).toContain(itemContentEditingClass);
-            basicComponent.editItemComponent.setKeepEditing(false);
-            dispatchMouseEvent(ageEditorElement.nativeElement, 'mouseleave');
+            basicComponent.editItemComponent.setEditing(false);
             fixture.detectChanges();
             expect(ageEditorElement.nativeElement.parentNode.classList).not.toContain(itemContentEditingClass);
         });
@@ -148,7 +150,7 @@ describe(`thy-properties`, () => {
 
     describe(`with column`, () => {
         let fixture: ComponentFixture<ThyPropertiesTestColumnComponent>;
-        let basicComponent: ThyPropertiesTestColumnComponent;
+        let testColumnComponent: ThyPropertiesTestColumnComponent;
 
         beforeEach(fakeAsync(() => {
             TestBed.configureTestingModule({
@@ -160,7 +162,7 @@ describe(`thy-properties`, () => {
 
         beforeEach(() => {
             fixture = TestBed.createComponent(ThyPropertiesTestColumnComponent);
-            basicComponent = fixture.debugElement.componentInstance;
+            testColumnComponent = fixture.debugElement.componentInstance;
             fixture.detectChanges();
         });
 
@@ -168,7 +170,28 @@ describe(`thy-properties`, () => {
             const trs = fixture.debugElement.queryAll(By.css('tr'));
             const tds = fixture.debugElement.queryAll(By.css('td'));
             expect(trs.length).toEqual(2);
-            expect(tds.length).toEqual(6);
+            expect(tds.length).toEqual(5);
+        });
+
+        it('should render item elements length eq property items length ', () => {
+            expect(testColumnComponent.propertiesComponent.items.length).toEqual(
+                testColumnComponent.propertiesComponent.itemElements.length
+            );
+        });
+
+        it('should set item span is work', () => {
+            let trs = fixture.debugElement.queryAll(By.css('tr'));
+            expect(trs.length).toEqual(2);
+            expect(trs[1].nativeElement.childElementCount).toEqual(2);
+            expect(trs[1].queryAll(By.css('td'))[1].attributes.colspan).toBe('2');
+            testColumnComponent.addressItemSpan = 3;
+            fixture.detectChanges();
+            trs = fixture.debugElement.queryAll(By.css('tr'));
+            expect(trs.length).toEqual(3);
+            expect(trs[1].nativeElement.childElementCount).toEqual(1);
+            expect(trs[1].queryAll(By.css('td'))[0].attributes.colspan).toBe('3');
+            expect(trs[2].nativeElement.childElementCount).toEqual(1);
+            expect(trs[2].queryAll(By.css('td'))[0].attributes.colspan).toBe('3');
         });
     });
 });