+++ src/select/examples/footer/footer.component.ts
@@ -1,17 +1,12 @@
 import { Component, OnInit } from '@angular/core';
+import { listOfOption } from '../mock-data';
 
 @Component({
     selector: 'thy-select-footer-example',
     templateUrl: './footer.component.html'
 })
 export class ThySelectFooterExampleComponent implements OnInit {
-    listOfOption: Array<{ label: string; value: string }> = [];
+    listOfOption = listOfOption;
 
-    ngOnInit() {
-        const children: Array<{ label: string; value: string }> = [];
-        for (let i = 10; i < 36; i++) {
-            children.push({ label: i.toString(36) + i, value: i.toString(36) + i });
-        }
-        this.listOfOption = children;
-    }
+    ngOnInit() {}
 }